Ansible installs ArgoCD
The rest of the config changes must come from ArgoCd Apps.

ANSIBLE
The playbook references the localhost, not as 127.0.0.1 but as localhost so a variable can be defined in the hostvars/localhost file.

The credentials (api_key) is obtained directly from the user_credentials file, which must contain a valid token for a user or
service account with cluster-admin privileges.  This is because the k8s_auth module has problems with the API endpoint and the default ingress controller using
certificates from different issuers.

The GitOps operator is installed from a role: argocd_install

The group_vars/all file, for lack of a better place contains the list of config changes that will be applied to the cluster.  Default values are false, for don't apply the change.

The host_vars/localhost defines the version of python to use on the manged servers.  This is probably something that only needs to be used in the test environment because it contains 2 different python3 versions.

Signed-off-by: GTPE Student <lab-user@bastion.5cvx2.internal>