Delete wait job after successful execution
If the post check job is successfull, the job is deleted.  This is because trying to apply the job again causes errors because some of the fields may be immutable

A PostSync hook is used, so the job is executed after the other resources are applied, in this case there is no need to define a different wave.

Signed-off-by: GTPE Student <lab-user@bastion.5cvx2.internal>